refactor: HomeFragmentをActivityに変更
