+++ android-project/MakaseteChoice/app/src/main/java/com/teambrack/makasetechoice/MainActivity.kt
@@ -1,16 +1,25 @@
 package com.teambrack.makasetechoice
 
 import android.os.Bundle
-import com.teambrack.makasetechoice.home.HomeFragment
+import androidx.databinding.DataBindingUtil
+import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.ViewModelProviders
+import com.teambrack.makasetechoice.databinding.ActivityMainBinding
+import com.teambrack.makasetechoice.home.HomeViewModel
 import dagger.android.support.DaggerAppCompatActivity
+import javax.inject.Inject
 
 class MainActivity : DaggerAppCompatActivity() {
+    @Inject
+    lateinit var viewModelFactory: ViewModelProvider.Factory
+    private val viewModel: HomeViewModel by lazy {
+        ViewModelProviders.of(this, viewModelFactory).get(HomeViewModel::class.java)
+    }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        supportFragmentManager.beginTransaction().apply {
-            replace(R.id.fragment_area, HomeFragment.newInstance(), HomeFragment.FRAGMENT_TAG)
-            commit()
+        DataBindingUtil.setContentView<ActivityMainBinding>(this, R.layout.activity_main).also{binding ->
+            binding.viewModel = viewModel
         }
     }
 }
